refactor(#zimic): improve location of unhandled request type
