+++ packages/zimic/src/interceptor/http/interceptor/types/requests.ts
@@ -1,10 +1,27 @@
+import { HttpBody } from '@/http';
 import { HttpSchema, HttpSchemaMethod, HttpSchemaPath } from '@/http/types/schema';
 import { Default } from '@/types/utils';
 
 import { HttpResponseFactoryContext } from '../../interceptorWorker/types/requests';
+import { HttpInterceptorRequest } from '../../requestHandler/types/requests';
 
 export type HttpInterceptorRequestContext<
   Schema extends HttpSchema,
   Method extends HttpSchemaMethod<Schema>,
   Path extends HttpSchemaPath<Schema, Method>,
 > = HttpResponseFactoryContext<Default<Default<Schema[Path][Method]>['request']>['body']>;
+
+export type UnhandledHttpInterceptorRequestPath = string;
+
+export type UnhandledHttpInterceptorRequestMethodSchema = HttpSchema.Method<{
+  request: {
+    headers: Record<string, string>;
+    searchParams: Record<string, string | string[]>;
+    body: HttpBody;
+  };
+}>;
+
+export type UnhandledHttpInterceptorRequest = Omit<
+  HttpInterceptorRequest<string, UnhandledHttpInterceptorRequestMethodSchema>,
+  'response'
+>;
